Реализованы ДТО и сервис под эндпоинты получения и изменения статусов запросов, добавлено исключение BadRequestException, исправлены ошибки в схеме БД, синхронизированы наименовая при сериализации/дессер

+++ main-server/src/main/java/main/server/request/RequestRepository.java
@@ -10,13 +10,13 @@ import java.util.Optional;
 public interface RequestRepository extends JpaRepository<ParticipationRequest, Long> {
     Optional<ParticipationRequest> findByEventIdAndRequesterId(Long eventId, Long requesterId);
 
-    int countByEventIdAndStatusEquals(Long eventId, RequestStatus requestStatus);
+    long countByEventIdAndStatusEquals(Long eventId, RequestStatus requestStatus);
 
     List<ParticipationRequest> findAllByRequesterId(Long requesterId);
 
     List<ParticipationRequest> findByEventId(Long eventId);
 
-    int countByIdInAndEventId(List<Long> requestIds, Long eventId);
+    long countByIdInAndEventId(List<Long> requestIds, Long eventId);
 
     List<ParticipationRequest> findAllByIdIn(List<Long> requestIds);
 }
